optimised the functionality to production standard

+++ Backend/controllers/deleteParticularFlag.js
@@ -1,5 +1,5 @@
 import redisClient from "../config/redisConfig.js";
-import { deleteFlag } from "../model/featureFlagsModel.js";
+import { deleteFlag , findWithId} from "../model/featureFlagsModel.js";
 
 export const deleteParticularFlag = async(req, res) =>{
     const id = parseInt(req.params?.id , 10 )//convert to integer in case string is received
@@ -7,20 +7,30 @@ export const deleteParticularFlag = async(req, res) =>{
         return res.status(400).json({ message: 'Invalid or missing ID' })
     }
     try{
-        // Soft delete: update is_active = false(since, analytics me use)
-        const result = await deleteFlag(id);
-        if(!result){
-            return res.status(400).json({message:`No flag with id -> {$id} is found or flag was already deactivated`})
+        //1.check if this id exists or not 
+        const flagStatus = await findWithId(id);
+        if(!flagStatus){
+            return res.status(400).json({message: `flag with id=${id} doesn't exist `})
         }
+        if(flagStatus.is_deleted){
+            return res.status(400).json({message: `flag with id=${id} was already deleted , no need to delete it again`})
+        }
+        // 2. Soft delete: update is_active = false(since, analytics me use)
+        const result = await deleteFlag(id);
+
+        //above checks already passed ,so , no definitely , result will not be empty 
         //delete stored instance of this flag from redis
-        const cacheKey = `flag:${id}`
-        await redisClient.del(cacheKey) //throws no error even if not present in redis
 
-        //also invalidate the flag:all (cache to store whole table)(since, a record deleted from table ), so -> invalidate already stored table in redis
-        const wholeTableCacheKey = `flag:all` ;
-        await redisClient.del(wholeTableCacheKey); //even if this cache not there in redis, still no error thrown
+        // 3. invalidate Redis Caches
+        const cacheKey = `flag:${id}`;
+        const wholeTableCacheKey = `flag:all`;
 
-        return res.status(200).json({message:`flag with id->${id} has been deactivated`, flag:result})
+        // using Promise.all for faster concurrent execution
+        await Promise.all([
+            redisClient.del(cacheKey),  //won't be giving any error even if cacheKey doesn't exist, feature of redis
+            redisClient.del(wholeTableCacheKey) //won't be giving any error even if wholeTAbleCacheKey doesn't exist, feature of redis
+        ]);
+        return res.status(200).json({message:`flag with id->${id} has been marked as deleted `, flag:result})
     }catch(err){
         console.log(err.stack)
         return res.status(500).json({error:'Flag deletion failed(backend issue) '})
